requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
google-auth==2.30.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
httplib2==0.22.0
idna==3.7
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
